refactor(shared-kernel): move PlatformRole from Identity
